+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New card "Ballus, Dogfight Enforcer Q"
 - New card "Split-Head Hydroturtle Q"
 - New card "Bladerush Skyterror Q"
+- New card "Ruthless Skyterror"
 
 ## [v2.3] - 09/04/2022